+++ tests/factories.py
@@ -10,5 +10,8 @@ class EncryptedTextFieldModelFactory(factory.DjangoModelFactory):
 
     digest_field = factory.Sequence('Text digest {}'.format)
     hmac_field = factory.Sequence('Text hmac {}'.format)
+
+    integer_pgp_pub_field = 42
     pgp_pub_field = factory.Sequence('Text with public key {}'.format)
+    integer_pgp_sym_field = 43
     pgp_sym_field = factory.Sequence('Text with symmetric key {}'.format)